requirements.txt
flask
ccxt
gunicorn
python-dotenv
